Splitted FooterHeading and FooterHeading, exported all Footer components that are needed to construct the story for the Footer implementation
