+++ packages/asc-core/package.json
@@ -14,8 +14,7 @@
   "version": "0.11.0",
   "main": "lib/index.js",
   "files": [
-    "lib",
-    "README.md"
+    "lib"
   ],
   "directories": {
     "lib": "lib",
