+++ packages/asc-ui/package.json
@@ -6,8 +6,7 @@
   "version": "0.12.0",
   "main": "lib/index.js",
   "files": [
-    "lib",
-    "README.md"
+    "lib"
   ],
   "typings": "lib/index.d.ts",
   "dependencies": {
