+++ packages/asc-ui/src/components/Footer/FooterBottom/index.ts
@@ -1,2 +1,2 @@
+export { default } from './FooterBottom'
 export { default as FooterBottomStyle } from './FooterBottomStyle'
-export { default as FooterBottom } from './FooterBottom'
