+++ packages/asc-ui/src/components/Footer/FooterContent/index.ts
@@ -1,2 +1,2 @@
+export { default } from './FooterContent'
 export { default as FooterContentStyle } from './FooterContentStyle'
-export { default as FooterContent } from './FooterContent'
