+++ packages/asc-ui/src/components/Footer/FooterHeading/FooterHeading.test.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react'
 import { ascDefaultTheme, ThemeProvider } from '@datapunt/asc-core'
 import { render, cleanup, fireEvent } from '@testing-library/react'
-import { FooterHeading, ToggleFooterHeader } from '.'
+import { FooterHeading, ToggleFooterHeader } from '../index'
 import { color } from '../../../utils'
 import 'jest-styled-components'
