+++ packages/asc-ui/src/components/Footer/FooterHeading/index.ts
@@ -1,3 +1,2 @@
-export { default as FooterHeading } from './FooterHeading'
-export { default as ToggleFooterHeader } from './ToggleFooterHeader'
-export { default as FooterToggle } from './FooterToggle'
+export { default } from './FooterHeading'
+export { default as FooterHeadingStyle } from './FooterHeadingStyle'
