+++ packages/asc-ui/src/components/Footer/FooterLinkList/FooterLinkList.test.tsx
@@ -1,8 +1,12 @@
 import * as React from 'react'
 import { render, cleanup } from '@testing-library/react'
 import { ascDefaultTheme, ThemeProvider } from '@datapunt/asc-core'
-import { FooterLinkList, FooterLinkListItem, FooterBottomLinkListItem } from '.'
-import FooterBottomLinkList from './FooterBottomLinkList'
+import {
+  FooterBottomLinkList,
+  FooterLinkList,
+  FooterLinkListItem,
+  FooterBottomLinkListItem,
+} from '../index'
 import { color } from '../../../utils'
 import 'jest-styled-components'
