+++ packages/asc-ui/src/components/Footer/FooterLinkList/index.tsx
@@ -1,4 +1,4 @@
-export { default as FooterLinkList } from './FooterLinkList'
+export { default } from './FooterLinkList'
 export { default as FooterLinkListItem } from './FooterLinkListItem'
 export { default as FooterBottomLinkList } from './FooterBottomLinkList'
 export { default as FooterBottomLinkListItem } from './FooterBottomLinkListItem'
