+++ packages/asc-ui/src/components/Footer/FooterToggle/index.ts
@@ -1,0 +1,2 @@
+export { default as ToggleFooterHeader } from './ToggleFooterHeader'
+export { default } from './FooterToggle'
