+++ packages/asc-ui/src/components/Footer/FooterTop/index.ts
@@ -1,2 +1,2 @@
+export { default } from './FooterTop'
 export { default as FooterTopStyle } from './FooterTopStyle'
-export { default as FooterTop } from './FooterTop'
