+++ packages/asc-ui/src/components/Footer/index.ts
@@ -1,4 +1,23 @@
-export { default as FooterStyle } from './FooterStyle'
-export { FooterTop } from './FooterTop'
-export { FooterBottom } from './FooterBottom'
-export { default as Footer } from './Footer'
+import { default as FooterBottom, FooterBottomStyle } from './FooterBottom'
+import { default as FooterContent, FooterContentStyle } from './FooterContent'
+import { default as FooterHeading, FooterHeadingStyle } from './FooterHeading'
+import { default as FooterStyle } from './FooterStyle'
+import { default as FooterTop, FooterTopStyle } from './FooterTop'
+
+export { default } from './Footer'
+export {
+  default as FooterLinkList,
+  FooterBottomLinkList,
+  FooterBottomLinkListItem,
+  FooterLinkListItem,
+} from './FooterLinkList'
+export { default as FooterToggle, ToggleFooterHeader } from './FooterToggle'
+export { FooterBottom, FooterContent, FooterHeading, FooterTop }
+
+export const FooterStyles: any = {
+  FooterBottomStyle,
+  FooterContentStyle,
+  FooterHeadingStyle,
+  FooterTopStyle,
+  FooterStyle,
+}
